+++ app/Http/Controllers/OperacionController.php
@@ -6,25 +6,46 @@ use App\Models\Mascota;
 use App\Models\Operacion;
 use App\Models\Veterinaria;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class OperacionController extends Controller
 {
     public function index($mascota_id)
     {
-        return view('vistas.operaciones.index',
-            [
-                'operaciones' => Operacion::where('mascota_id', '=', $mascota_id)->orderBy('id', 'asc')->get(),
-                'mascota' => Mascota::findOrFail($mascota_id),
-            ]);
+        $mascota = Mascota::findOrFail($mascota_id);
+        if ($mascota->usuario_id == Auth::id() || Auth::user()->admin){
+            if (Auth::user()->admin) {
+                $operaciones = Operacion::where('mascota_id','=',$mascota_id)
+                    ->orderBy('id', 'asc')
+                    ->get();
+            } else {
+                $operaciones = Operacion::where('mascota_id','=',$mascota_id)
+                    ->where('usuario_id','=', Auth::id())
+                    ->orderBy('id', 'asc')
+                    ->get();
+            }
+
+            return view('vistas.operaciones.index',
+                [
+                    'operaciones' => $operaciones,
+                    'mascota' => Mascota::findOrFail($mascota_id),
+                ]);
+        }
+        return redirect('mascotas');
+
     }
 
     public function create($mascota_id)
     {
-        return view('vistas.operaciones.create',
-            [
-                'mascota_id' => $mascota_id,
-                'veterinarias' => Veterinaria::all(),
-            ]);
+        $mascota = Mascota::findOrFail($mascota_id);
+        if ($mascota->usuario_id == Auth::id() || Auth::user()->admin){
+            return view('vistas.operaciones.create',
+                [
+                    'mascota_id' => $mascota_id,
+                    'veterinarias' => Veterinaria::all(),
+                ]);
+        }
+        return redirect('mascotas');
     }
 
     public function store($mascota_id, Request $request)
@@ -34,6 +55,7 @@ class OperacionController extends Controller
         $operacion->fecha = $request['fecha'];
         $operacion->mascota_id = $mascota_id;
         $operacion->veterinaria_id = $request['veterinaria_id'];
+        $operacion->usuario_id = Auth::id();
         $operacion->save();
 
         return redirect('mascotas/'.$mascota_id.'/operaciones');
@@ -41,12 +63,16 @@ class OperacionController extends Controller
 
     public function edit($mascota_id, $id)
     {
-        return view('vistas.operaciones.edit',
-            [
-                'operacion' => Operacion::findOrFail($id),
-                'mascota_id' => $mascota_id,
-                'veterinarias' => Veterinaria::all(),
-            ]);
+        $operacion = Operacion::findOrFail($id);
+        if ($operacion->usuario_id == Auth::id() || Auth::user()->admin) {
+            return view('vistas.operaciones.edit',
+                [
+                    'operacion' => $operacion,
+                    'mascota_id' => $mascota_id,
+                    'veterinarias' => Veterinaria::all(),
+                ]);
+        }
+        return redirect('mascotas');
     }
 
     public function update($mascota_id, Request $request, $id)
@@ -63,8 +89,10 @@ class OperacionController extends Controller
     public function destroy($mascota_id, $id)
     {
         $operacion = Operacion::findOrFail($id);
-        $operacion->delete();
-
-        return redirect('mascotas/'.$mascota_id.'/operaciones');
+        if ($operacion->usuario_id == Auth::id() || Auth::user()->admin){
+            $operacion->delete();
+            return redirect('mascotas/'.$mascota_id.'/operaciones');
+        }
+        return redirect('mascotas');
     }
 }
